server/requirements.txt
amqp==5.0.2
asgiref==3.3.1
beautifulsoup4==4.9.3
bert-extractive-summarizer==0.6.1
billiard==3.6.3.0
blis==0.7.4
bokeh==2.2.3
catalogue==1.0.0
celery==5.0.5
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cycler==0.10.0
cymem==2.0.5
decorator==4.4.2
dj-rest-auth==2.1.2
Django==3.1.4
django-celery-beat==2.1.0
django-contrib-postgres==0.0.1
django-cors-headers==3.6.0
django-extensions==3.1.0
django-filter==2.4.0
django-organizations==2.0.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
drf-nested-routers==0.92.5
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.3.1/en_core_web_md-2.3.1.tar.gz
filelock==3.0.12
httpobs-cli==1.0.2
idna==2.10
inscriptis==1.1.1
jellyfish==0.8.2
Jinja2==2.11.2
joblib==1.0.0
kiwisolver==1.3.1
kombu==5.0.2
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.3
murmurhash==1.0.5
networkx==2.5
numpy==1.19.4
packaging==20.8
pandas==1.2.0
Pillow==8.1.0
plac==1.1.3
preshed==3.0.5
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
regex==2020.11.13
requests==2.24.0
sacremoses==0.0.43
scikit-learn==0.24.0
scipy==1.6.0
seaborn==0.11.1
segtok==1.5.10
six==1.15.0
soupsieve==2.1
spacy==2.3.5
sqlparse==0.4.1
srsly==1.0.5
ssh-audit==2.3.1
tabulate==0.8.7
thinc==7.4.5
threadpoolctl==2.1.0
tokenizers==0.9.4
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tornado==6.1
tqdm==4.55.1
transformers==4.1.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
wasabi==0.8.0
wcwidth==0.2.5
yake @ git+https://github.com/LIAAD/yake@481b684ade75b67027e63c5de849de19f198d96c
